% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_dw_vars.R
\name{append_dw_vars}
\alias{append_dw_vars}
\title{Conveniently append common 'deweathering' variables to an air quality time
series}
\usage{
append_dw_vars(
  data,
  vars = c("trend", "hour", "weekday", "weekend", "yday", "week", "month"),
  abbr = TRUE,
  ...,
  .date = "date"
)
}
\arguments{
\item{data}{An input \code{data.frame} with at least one date(time) column.}

\item{vars}{A character vector of variables of interest. Possible options
include:
- \code{"trend"}: a numeric expression of the overall time series
- \code{"hour"}: the hour of the day (0-23)
- \code{"weekday"}: the day of the week (Sunday through Saturday)
- \code{"weekend"}: whether it is a weekend (Saturday, Sunday) or weekday
- \code{"yday"}: the day of the year
- \code{"week"}: the week of the year
- \code{"month"}: the month of the year}

\item{abbr}{Abbreviate weekday and month strings? Defaults to \code{TRUE}, which
tends to look better in plots.}

\item{...}{Not used}

\item{.date}{The name of the 'date' column to use for manipulation.}
}
\description{
This function conveniently manipulates a datetime ('POSIXct') column (by
default named 'date') into a series of columns which are useful features in
deweather models. Used internally by \code{\link[=build_dw_model]{build_dw_model()}} and
\code{\link[=tune_dw_model]{tune_dw_model()}}, but can be used directly by users if desired.
}
\seealso{
\code{\link[openair:cutData]{openair::cutData()}} for more flexible time series data conditioning.
}
