% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqroadside.R
\docType{data}
\name{aqroadside}
\alias{aqroadside}
\title{Example air quality monitoring data for openair}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 149040 rows and 11 columns.
}
\usage{
aqroadside
}
\description{
\describe{
\item{date}{Observation date/time stamp in year-month-day hour:minute:second
format (POSIXct).}
\item{ws}{Wind speed, in m/s, as numeric vector.}
\item{wd}{Wind direction, in degrees from North, as a numeric vector.}
\item{nox}{Oxides of nitrogen concentration, in ppb, as a numeric vector.}
\item{no2}{Nitrogen dioxide concentration, in ppb, as a numeric vector.}
\item{o3}{Ozone concentration, in ppb, as a numeric vector.}
\item{pm10}{Particulate PM10 fraction measurement, in ug/m3 (raw TEOM), as a
numeric vector.}
\item{so2}{Sulfur dioxide concentration, in ppb, as a numeric vector.}
\item{co}{Carbon monoxide concentration, in ppm, as a numeric vector.}
\item{pm25}{Particulate PM2.5 fraction measurement, in ug/m3, as a numeric
vector.}
}
}
\examples{
# basic structure
head(aqroadside)
}
\keyword{datasets}
