% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dw_model.R
\name{build_dw_model}
\alias{build_dw_model}
\title{Build a Deweather Model}
\usage{
build_dw_model(
  data,
  pollutant,
  vars = c("trend", "ws", "wd", "hour", "weekday", "air_temp"),
  tree_depth = 5,
  trees = 200L,
  learn_rate = 0.1,
  mtry = NULL,
  min_n = 10L,
  loss_reduction = 0,
  sample_size = 1L,
  stop_iter = 190L,
  engine = c("xgboost", "lightgbm"),
  ...,
  .date = "date"
)
}
\arguments{
\item{data}{An input \code{data.frame} containing one pollutant column (defined
using \code{pollutant}) and a collection of feature columns (defined using
\code{vars}).}

\item{pollutant}{The name of the column (likely a pollutant) in \code{data} to
predict.}

\item{vars}{The name of the columns in \code{data} to use as model features -
i.e., to predict the values in the \code{pollutant} column. Any character
columns will be coerced to factors. \code{"hour"}, \code{"weekday"}, \code{"trend"},
\code{"yday"}, \code{"week"}, and \code{"month"} are special terms and will be passed to
\code{\link[=append_dw_vars]{append_dw_vars()}} if not present in \code{names(data)}.}

\item{tree_depth}{An integer for the maximum depth of the tree (i.e. number
of splits) (specific engines only).}

\item{trees}{An integer for the number of trees contained in
the ensemble.}

\item{learn_rate}{A number for the rate at which the boosting algorithm adapts
from iteration-to-iteration (specific engines only). This is sometimes referred to
as the shrinkage parameter.}

\item{mtry}{A number for the number (or proportion) of predictors that will
be randomly sampled at each split when creating the tree models
(specific engines only).}

\item{min_n}{An integer for the minimum number of data points
in a node that is required for the node to be split further.}

\item{loss_reduction}{A number for the reduction in the loss function required
to split further (specific engines only).}

\item{sample_size}{A number for the number (or proportion) of data that is
exposed to the fitting routine. For \code{xgboost}, the sampling is done at
each iteration while \code{C5.0} samples once during training.}

\item{stop_iter}{The number of iterations without improvement before
stopping (specific engines only).}

\item{engine}{A single character string specifying what computational engine
to use for fitting.}

\item{...}{Not current used.}

\item{.date}{The name of the 'date' column which defines the air quality
timeseries. Passed to \code{\link[=append_dw_vars]{append_dw_vars()}} if needed. Also used to extract
the time zone of the data for later restoration if \code{trend} is used as a
variable.}
}
\value{
a 'Deweather' object for further analysis
}
\description{
This function builds a boosted decision tree machine learning model with
useful methods for interrogating it in an air quality and meteorological
context. Currently, only the \link[xgboost:xgboost]{xgboost} engine is
supported.
}
