% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting.R
\name{buildMod}
\alias{buildMod}
\title{Function to apply meteorological normalisation to.}
\usage{
buildMod(dat, vars = c("trend", "ws", "wd", "hour", "weekday", "temp"),
  pollutant = "nox", B = 100, n.core = 4)
}
\arguments{
\item{dat}{Data frame to analyse. Must contain a POSIXct field
called \code{date}.}

\item{vars}{Explanatory variables to use. These variables will be
used to build the gbm model. Note that the model must include
a trend component. Several variables can be automatically
calculated (see \code{\link{prepData}} for details).}

\item{pollutant}{The name of the variable to apply meteorological
normalisation to.}

\item{B}{Number of bootstrap simulations for partial dependence
plots.}

\item{n.core}{Number of cores to use for parallel processing.}
}
\value{
Returns a list including the model, influence data frame
    and partial dependence data frame.
}
\description{
Function to do meteorological normalisation
}
\details{
This is the main function to apply a gbm model to a data set.
}
\author{
David Carslaw
}

