% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boosting.R
\name{deMet}
\alias{deMet}
\title{Function to apply meteorological normalisation to.}
\usage{
deMet(dat, vars = c("ws", "wd"), pollutant = "nox")
}
\arguments{
\item{dat}{Data frame to analyse.}

\item{vars}{Explanatory variables to use}

\item{pollutant}{The name of the variable to apply meteorological
normalisation to.}
}
\value{
A list of stuff
}
\description{
Function to do meteorological normalisation
}
\details{
Info here
}
\author{
David Carslaw
}

