% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deweather-package.R
\docType{package}
\name{deweather-package}
\alias{deweather}
\alias{deweather-package}
\title{deweather: Remove the influence of weather on air quality data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Model and account for (or remove) the effect of meteorology on atmospheric composition data. The technique uses boosted regression trees via the tidymodels framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openair-project.github.io/deweather/}
  \item \url{https://github.com/openair-project/deweather}
  \item Report bugs at \url{https://github.com/openair-project/deweather/issues}
}

}
\author{
\strong{Maintainer}: David Carslaw \email{david.carslaw@york.ac.uk} (\href{https://orcid.org/0000-0003-0991-950X}{ORCID})

Authors:
\itemize{
  \item Jack Davison \email{jack.davison@ricardo.com} (\href{https://orcid.org/0000-0003-2653-6615}{ORCID})
}

}
\keyword{internal}
