% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diurnalGbm.R
\name{diurnalGbm}
\alias{diurnalGbm}
\title{Plot diurnal changes, removing the effect of meteorology}
\usage{
diurnalGbm(
  input_data,
  vars = c("ws", "wd", "hour", "weekday"),
  pollutant = "nox",
  dates = c("01/01/2012", "31/12/2012", "31/12/2013"),
  ylab = "value",
  plot = TRUE
)
}
\arguments{
\item{input_data}{A data frame to analyse.}

\item{vars}{The explanatory variables used in the model.}

\item{pollutant}{Name of the pollutant to apply meteorological normalisation
to.}

\item{dates}{A vector of dates, length three. These dates are used to
partition the data into two categories (before/after). The date format is
UK e.g. \code{date = c("19/2/2005", "19/2/2007", "19/2/2010")}.}

\item{ylab}{Label for y-axis.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}
}
\value{
Some data
}
\description{
This function calculates the diurnal profile of a pollutant with the effect
of meteorology removed. Its primary use is to compare two periods to
determine whether there has been a shift in diurnal profile e.g. due to some
intervention.
}
\author{
David Carslaw
}
