% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_tdw_model.R
\name{finalise_tdw_model}
\alias{finalise_tdw_model}
\title{Use the 'best parameters' determined by \code{\link[=tune_dw_model]{tune_dw_model()}} to build a
Deweather Model}
\usage{
finalise_tdw_model(tdw, data, params = list(), ..., .date = "date")
}
\arguments{
\item{tdw}{A deweather tuning object created with \code{\link[=tune_dw_model]{tune_dw_model()}}.}

\item{data}{An input \code{data.frame} containing one pollutant column (defined
using \code{pollutant}) and a collection of feature columns (defined using
\code{vars}). This must be provided in addition to \code{tdw} as it is expected most
users will have provided \code{\link[=tune_dw_model]{tune_dw_model()}} with a sampled dataset.}

\item{params}{A named list. These parameters are used to override the
\code{best_params} defined within \code{tdw}. For example, if the 'best' parameter
for \code{trees} is 50, \code{params = list(trees = 100)} will set it to 100 instead.}

\item{...}{Not current used.}

\item{.date}{The name of the 'date' column which defines the air quality
timeseries. Passed to \code{\link[=append_dw_vars]{append_dw_vars()}} if needed. Also used to extract
the time zone of the data for later restoration if \code{trend} is used as a
variable.}
}
\description{
This function takes the output of \code{\link[=tune_dw_model]{tune_dw_model()}} and uses the
\code{best_params} defined within. This is a convenient wrapper around
\code{\link[=build_dw_model]{build_dw_model()}} if you have already run \code{\link[=tune_dw_model]{tune_dw_model()}} and are broadly
happy with the parameters it has chosen. That being said, the \code{params}
argument can be used to override specific hyperparameters.
}
\seealso{
\code{\link[=build_dw_model]{build_dw_model()}}
}
\author{
Jack Davison
}
