% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{gbmInf}
\alias{gbmInf}
\title{Plot a GBM influence plot}
\usage{
gbmInf(dat)
}
\arguments{
\item{dat}{Model object from running \code{\link[=buildMod]{buildMod()}}.}
}
\value{
Plot
}
\description{
Plot a GBM influence plot
}
\author{
David Carslaw
}
