% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbmInteractions.R
\name{gbmInteractions}
\alias{gbmInteractions}
\title{Quantify most important 2-way interactions}
\usage{
gbmInteractions(dw_model)
}
\arguments{
\item{dw_model}{Model object from running \code{\link[=buildMod]{buildMod()}}.}
}
\value{
Interaction information.
}
\description{
Quantify most important 2-way interactions
}
\author{
David Carslaw
}
