% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metSim.R
\name{get_constrained_random_index_rcpp}
\alias{get_constrained_random_index_rcpp}
\title{Function to shuffle dates}
\usage{
get_constrained_random_index_rcpp(dates, day_window = 15, hour_window = 2)
}
\arguments{
\item{dates}{Dates to shuffle.}
}
\description{
Function to shuffle dates
}
