% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dw.R
\name{get_dw_pollutant}
\alias{get_dw_pollutant}
\alias{get_dw_vars}
\alias{get_dw_engine}
\alias{get_dw_params}
\alias{get_dw_input_data}
\alias{get_dw_model}
\alias{get_dw_importance}
\title{Getters for various deweather model features}
\usage{
get_dw_pollutant(dw)

get_dw_vars(dw)

get_dw_engine(dw)

get_dw_params(dw, param = NULL)

get_dw_input_data(dw)

get_dw_model(dw)

get_dw_importance(dw, aggregate_factors = FALSE, sort = TRUE)
}
\arguments{
\item{dw}{A deweather model created with \code{\link[=build_dw_model]{build_dw_model()}}.}

\item{param}{For \code{\link[=get_dw_params]{get_dw_params()}} and \code{\link[=get_tdw_best_params]{get_tdw_best_params()}}. The default
(\code{NULL}) returns a list of model parameters. \code{param} will return one
specific parameter as a character vector.}

\item{aggregate_factors}{Defaults to \code{FALSE}. If \code{TRUE}, the importance of
factor inputs (e.g., Weekday) will be summed into a single variable. This
only applies to certain engines (e.g., \code{"xgboost"}) which report factor
importance as disaggregate features.}

\item{sort}{If \code{TRUE}, the default, features will be sorted by their
importance. If \code{FALSE}, they will be sorted alphabetically. In
\code{\link[=plot_dw_importance]{plot_dw_importance()}} this will change the ordering of the y-axis, whereas
in \code{\link[=get_dw_importance]{get_dw_importance()}} it will change whether \code{var} is returned as a
factor or character data type.}
}
\value{
Typically a character vector, except:
\itemize{
\item \code{\link[=get_dw_params]{get_dw_params()}}: a list, unless \code{param} is set.
\item \code{\link[=get_dw_importance]{get_dw_importance()}}: a \code{data.frame}
\item \code{\link[=get_dw_model]{get_dw_model()}}: A \link[parsnip:model_fit]{parsnip::model_fit} object
}
}
\description{
\code{deweather} provides multiple 'getter' functions for extracting relevant
model features from a deweather model and/or tuning objects. These are a
useful convenience, particularly in conjunction with R's \link[=pipeOp]{pipe}
operator (\verb{|>}).
}
\seealso{
Other Object 'Getter' Functions: 
\code{\link{get_tdw_pollutant}()}
}
\concept{Object 'Getter' Functions}
