% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2Way.R
\name{plot2Way}
\alias{plot2Way}
\title{Plot two-way interactions from gbm model}
\usage{
plot2Way(
  dw_model,
  variable = c("ws", "air_temp"),
  res = 100,
  exclude = TRUE,
  cols = "default",
  dist = 0.05,
  plot = TRUE,
  ...
)
}
\arguments{
\item{dw_model}{Model object from running \code{\link[=buildMod]{buildMod()}}.}

\item{variable}{The variables to plot. Must be of length two e.g. \code{variables = c("ws", "wd")}.}

\item{res}{Resolution in x-y, i.e., number of points in each dimension.}

\item{exclude}{Should surfaces exclude predictions too far from original
data? The default is \code{TRUE}.}

\item{cols}{Colours to be used for plotting, passed to
\code{\link[openair:openColours]{openair::openColours()}}.}

\item{dist}{When plotting surfaces, \code{dist} controls how far from the original
data the predictions should be made. See \code{exclude.too.far} from the \code{mgcv}
package. Data are first transformed to a unit square. Values should be
between \code{0} and \code{1}.}

\item{plot}{Should a plot be produced? \code{FALSE} can be useful when analysing
data to extract plot components and plotting them in other ways.}

\item{...}{Other arguments to be passed for plotting.}
}
\value{
To add
}
\description{
Plot two-way interactions from gbm model
}
\seealso{
Other deweather model plotting functions: 
\code{\link{plotInfluence}()},
\code{\link{plotPD}()}
}
\author{
David Carslaw
}
\concept{deweather model plotting functions}
