% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dw_importance.R
\name{plot_dw_importance}
\alias{plot_dw_importance}
\title{Visualise deweather model feature importance}
\usage{
plot_dw_importance(
  dw,
  aggregate_factors = FALSE,
  sort = TRUE,
  cols = "tol",
  ...,
  .plot = TRUE,
  .plot_engine = NULL
)
}
\arguments{
\item{dw}{A deweather model created with \code{\link[=build_dw_model]{build_dw_model()}}.}

\item{aggregate_factors}{Defaults to \code{FALSE}. If \code{TRUE}, the importance of
factor inputs (e.g., Weekday) will be summed into a single variable. This
only applies to certain engines (e.g., \code{"xgboost"}) which report factor
importance as disaggregate features.}

\item{sort}{If \code{TRUE}, the default, features will be sorted by their
importance. If \code{FALSE}, they will be sorted alphabetically. In
\code{\link[=plot_dw_importance]{plot_dw_importance()}} this will change the ordering of the y-axis, whereas
in \code{\link[=get_dw_importance]{get_dw_importance()}} it will change whether \code{var} is returned as a
factor or character data type.}

\item{cols}{Colours to use for plotting. See \code{\link[openair:openColours]{openair::openColours()}}.}

\item{...}{Not currently used.}

\item{.plot}{When \code{FALSE}, return a \code{data.frame} of plot data instead of a
plot.}

\item{.plot_engine}{The plotting engine to use. One of \code{"ggplot2"}, which
returns a static plot, or \code{"plotly"}, which returns a dynamic HTML plot.}
}
\value{
a \link[ggplot2:ggplot2-package]{ggplot2} figure
}
\description{
Visualise the feature importance (\% Gain for boosted tree models, permutation
importance for random forest models) for each variable of a deweather model
as a bar chart, with some customisation.
}
\section{Plotting Engines}{


\code{deweather} offers different plotting engines for different purposes. At
the moment, two plotting engines are supported:
\itemize{
\item \code{"ggplot2"}, for static plotting. This engine produces plots which can
be easily saved to a \code{.png}, \code{.svg}, or other 'static' file format. To save
a \code{ggplot2} plot, it is recommended to use the \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
function.
\item \code{"plotly"}, for dynamic plotting. This engine produces HTML plots which
are suitable for embedding into \code{quarto} or \code{rmarkdown} documents, or for
use in \code{shiny} applications. These can be saved using functions like
\code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}.
}

The plotting engine is defined using the \code{.plot_engine} argument in any
\verb{plot_*} function in \code{deweather}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# use default
plot_dw_importance(dw)

# set to ggplot2 (static)
plot_dw_importance(dw, .plot_engine = "ggplot2")

# set to plotly (HTML)
plot_dw_importance(dw, .plot_engine = "plotly")
}\if{html}{\out{</div>}}

When \code{.plot_engine} is not set, the engine defaults to \code{"ggplot2"}.
However, this option can be overridden by using the \code{deweather.plot_engine}
global option.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set once per session
options("deweather.plot_engine" = "plotly")

# now defaults to "plotly" over "ggplot2"
plot_dw_importance(dw)
}\if{html}{\out{</div>}}

Note that not all arguments in a function may apply to all plotting
engines. For example, hexagonal binning in \code{\link[=plot_tdw_testing_scatter]{plot_tdw_testing_scatter()}} is
supported in \code{ggplot2} but not in \code{plotly} at time of writing.
}

\examples{
\dontrun{
dw <- build_dw_model(aqroadside, "no2")
plot_dw_importance(dw)
}

}
\author{
Jack Davison
}
