% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dw_shap_importance.R
\name{plot_dw_shap_importance}
\alias{plot_dw_shap_importance}
\title{Plot SHAP importance for Deweather models}
\usage{
plot_dw_shap_importance(
  dw,
  vars = NULL,
  n = NULL,
  prop = 0.05,
  nsim = 50,
  cols = "turbo",
  ...,
  .plot = TRUE,
  .plot_engine = NULL
)
}
\arguments{
\item{dw}{A deweather model created with \code{\link[=build_dw_model]{build_dw_model()}}.}

\item{vars}{Character. The name of the variable(s) to plot. Must be one of
the variables used in the model. If \code{NULL}, all variables will be plotted
in order of importance.}

\item{n}{The number of observations to use for calculating SHAP values. If
\code{NULL} (default), uses \code{prop} to determine the sample size.}

\item{prop}{The proportion of input data to use for calculating the partial
dependence profile, between 0 and 1. Default is \code{0.05} (5\% of input data).
Ignored if \code{n} is specified.}

\item{nsim}{The number of Monte Carlo repetitions to use for estimating each
Shapley value, for engines which require \code{fastshap::explain()} (currently
\code{"ranger"}). Ignored for other engines.}

\item{cols}{Colours to use for plotting. See \code{\link[openair:openColours]{openair::openColours()}}.}

\item{...}{Not currently used.}

\item{.plot}{When \code{FALSE}, return a \code{data.frame} of plot data instead of a
plot.}

\item{.plot_engine}{The plotting engine to use. One of \code{"ggplot2"}, which
returns a static plot, or \code{"plotly"}, which returns a dynamic HTML plot.}
}
\value{
A \code{ggplot2} or \code{plotly} object showing the SHAP importance of the
specified features. If \code{.plot} is \code{FALSE}, returns a list containing the
data used for plotting and a \code{shapviz} object.
}
\description{
This function computes SHAP values for a Deweather model and plots the
importance of specified features. SHAP values show the contribution of each
feature to the model's predictions, allowing for a better understanding of
which features are most influential in the model. Wider ranges of SHAP values
indicate greater importance of the feature in the model's predictions. The
colour represents the normalised feature value, with darker colours
indicating higher values of the feature. This allows for the identification
of patterns in how feature values relate to their importance in the model's
predictions.
}
\section{Plotting Engines}{


\code{deweather} offers different plotting engines for different purposes. At
the moment, two plotting engines are supported:
\itemize{
\item \code{"ggplot2"}, for static plotting. This engine produces plots which can
be easily saved to a \code{.png}, \code{.svg}, or other 'static' file format. To save
a \code{ggplot2} plot, it is recommended to use the \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
function.
\item \code{"plotly"}, for dynamic plotting. This engine produces HTML plots which
are suitable for embedding into \code{quarto} or \code{rmarkdown} documents, or for
use in \code{shiny} applications. These can be saved using functions like
\code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}.
}

The plotting engine is defined using the \code{.plot_engine} argument in any
\verb{plot_*} function in \code{deweather}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# use default
plot_dw_importance(dw)

# set to ggplot2 (static)
plot_dw_importance(dw, .plot_engine = "ggplot2")

# set to plotly (HTML)
plot_dw_importance(dw, .plot_engine = "plotly")
}\if{html}{\out{</div>}}

When \code{.plot_engine} is not set, the engine defaults to \code{"ggplot2"}.
However, this option can be overridden by using the \code{deweather.plot_engine}
global option.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set once per session
options("deweather.plot_engine" = "plotly")

# now defaults to "plotly" over "ggplot2"
plot_dw_importance(dw)
}\if{html}{\out{</div>}}

Note that not all arguments in a function may apply to all plotting
engines. For example, hexagonal binning in \code{\link[=plot_tdw_testing_scatter]{plot_tdw_testing_scatter()}} is
supported in \code{ggplot2} but not in \code{plotly} at time of writing.
}

\examples{
\dontrun{
dw <- build_dw_model(aqroadside, "no2")
plot_dw_shap_importance(dw)
}
}
\author{
Jack Davison
}
