% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_trend.R
\name{plot_sim_trend}
\alias{plot_sim_trend}
\title{Plot a simulated 'deweathered' trend, optionally with its input data}
\usage{
plot_sim_trend(
  sim,
  dw = NULL,
  avg.time = NULL,
  ylim = NULL,
  names = c("Met Simulation", "Original Data"),
  cols = "tol",
  ...,
  .plot = TRUE,
  .plot_engine = NULL
)
}
\arguments{
\item{sim}{The output of \code{\link[=simulate_dw_met]{simulate_dw_met()}}; a \code{data.frame} with first
column \code{date} and second column a numeric pollutant.}

\item{dw}{Optionally, the input model used to construct \code{sim}. If provided,
the original timeseries will be plotted alongside the simulated trend for
easy comparison.}

\item{avg.time}{Passed to \code{\link[openair:timeAverage]{openair::timeAverage()}}.}

\item{ylim}{The limits of the y-axis.}

\item{names}{A character vector of length two, used to label the simulated
dataset and the original dataset.}

\item{cols}{Colours to use for plotting. See \code{\link[openair:openColours]{openair::openColours()}}.}

\item{...}{Not currently used.}

\item{.plot}{When \code{FALSE}, return a \code{data.frame} of plot data instead of a
plot.}

\item{.plot_engine}{The plotting engine to use. One of \code{"ggplot2"}, which
returns a static plot, or \code{"plotly"}, which returns a dynamic HTML plot.}
}
\description{
This function conveniently plots the trend from a simulated deweathered time
series, with the option to overlay the trend from the original input data.
The data can also be averaged over specified time intervals for clearer
visualisation.
}
\section{Plotting Engines}{


\code{deweather} offers different plotting engines for different purposes. At
the moment, two plotting engines are supported:
\itemize{
\item \code{"ggplot2"}, for static plotting. This engine produces plots which can
be easily saved to a \code{.png}, \code{.svg}, or other 'static' file format. To save
a \code{ggplot2} plot, it is recommended to use the \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
function.
\item \code{"plotly"}, for dynamic plotting. This engine produces HTML plots which
are suitable for embedding into \code{quarto} or \code{rmarkdown} documents, or for
use in \code{shiny} applications. These can be saved using functions like
\code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}.
}

The plotting engine is defined using the \code{.plot_engine} argument in any
\verb{plot_*} function in \code{deweather}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# use default
plot_dw_importance(dw)

# set to ggplot2 (static)
plot_dw_importance(dw, .plot_engine = "ggplot2")

# set to plotly (HTML)
plot_dw_importance(dw, .plot_engine = "plotly")
}\if{html}{\out{</div>}}

When \code{.plot_engine} is not set, the engine defaults to \code{"ggplot2"}.
However, this option can be overridden by using the \code{deweather.plot_engine}
global option.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set once per session
options("deweather.plot_engine" = "plotly")

# now defaults to "plotly" over "ggplot2"
plot_dw_importance(dw)
}\if{html}{\out{</div>}}

Note that not all arguments in a function may apply to all plotting
engines. For example, hexagonal binning in \code{\link[=plot_tdw_testing_scatter]{plot_tdw_testing_scatter()}} is
supported in \code{ggplot2} but not in \code{plotly} at time of writing.
}

\author{
Jack Davison
}
