% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tdw_tuning_metrics.R
\name{plot_tdw_tuning_metrics}
\alias{plot_tdw_tuning_metrics}
\title{Plot Tuning Metrics from \code{\link[=tune_dw_model]{tune_dw_model()}}}
\usage{
plot_tdw_tuning_metrics(
  tdw,
  x = NULL,
  group = NULL,
  facet = NULL,
  show_std_err = TRUE,
  cols = "tol",
  ...,
  .plot = TRUE,
  .plot_engine = NULL
)
}
\arguments{
\item{tdw}{A deweather tuning object created with \code{\link[=tune_dw_model]{tune_dw_model()}}.}

\item{x}{The tuned parameter to plot on the x-axis. If not selected, the
first parameter in the \code{metrics} dataset will be chosen.}

\item{group, facet}{Additional tuned parameters other than \code{x}, used to
further control the plot. \code{group} colours the plot by another parameter,
and \code{facet} splits the diagram into additional panels. Neither \code{group} nor
\code{facet} can be the same parameter as \code{x}.}

\item{show_std_err}{Show the standard error using error bars?}

\item{cols}{Colours to use for plotting. See \code{\link[openair:openColours]{openair::openColours()}}.}

\item{...}{Not currently used.}

\item{.plot}{When \code{FALSE}, return a \code{data.frame} of plot data instead of a
plot.}

\item{.plot_engine}{The plotting engine to use. One of \code{"ggplot2"}, which
returns a static plot, or \code{"plotly"}, which returns a dynamic HTML plot.}
}
\description{
This function creates a plot of the tuning metrics from a \code{TuneDeweather}
object created using \code{\link[=tune_dw_model]{tune_dw_model()}}. It visualises how different
hyperparameter values affect model performance (RMSE and RSQ). This allows
for the 'best' parameters to be refined through visual inspection. This plot
is likely most effective with between 1 and 3 simultaneously tuned
parameters; any more will impede plot interpretation.
}
\section{Plotting Engines}{


\code{deweather} offers different plotting engines for different purposes. At
the moment, two plotting engines are supported:
\itemize{
\item \code{"ggplot2"}, for static plotting. This engine produces plots which can
be easily saved to a \code{.png}, \code{.svg}, or other 'static' file format. To save
a \code{ggplot2} plot, it is recommended to use the \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}
function.
\item \code{"plotly"}, for dynamic plotting. This engine produces HTML plots which
are suitable for embedding into \code{quarto} or \code{rmarkdown} documents, or for
use in \code{shiny} applications. These can be saved using functions like
\code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}.
}

The plotting engine is defined using the \code{.plot_engine} argument in any
\verb{plot_*} function in \code{deweather}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# use default
plot_dw_importance(dw)

# set to ggplot2 (static)
plot_dw_importance(dw, .plot_engine = "ggplot2")

# set to plotly (HTML)
plot_dw_importance(dw, .plot_engine = "plotly")
}\if{html}{\out{</div>}}

When \code{.plot_engine} is not set, the engine defaults to \code{"ggplot2"}.
However, this option can be overridden by using the \code{deweather.plot_engine}
global option.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# set once per session
options("deweather.plot_engine" = "plotly")

# now defaults to "plotly" over "ggplot2"
plot_dw_importance(dw)
}\if{html}{\out{</div>}}

Note that not all arguments in a function may apply to all plotting
engines. For example, hexagonal binning in \code{\link[=plot_tdw_testing_scatter]{plot_tdw_testing_scatter()}} is
supported in \code{ggplot2} but not in \code{plotly} at time of writing.
}

\examples{
\dontrun{
tdw <- tune_dw_model(aqroadside, "no2", trees = c(1, 5))
plot_tdw_tuning_metrics(tdw)
}

}
\seealso{
Other Model Tuning Functions: 
\code{\link{plot_tdw_testing_scatter}()},
\code{\link{tune_dw_model}()}
}
\author{
Jack Davison
}
\concept{Model Tuning Functions}
