% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInfluence.R
\name{plotInfluence}
\alias{plotInfluence}
\title{Plot a GBM influence plot}
\usage{
plotInfluence(dw_model)
}
\arguments{
\item{dw_model}{Model object from running \code{\link[=buildMod]{buildMod()}}.}
}
\value{
Plot
}
\description{
Plot a GBM influence plot
}
\seealso{
Other deweather model plotting functions: 
\code{\link{plot2Way}()},
\code{\link{plotAllPD}()},
\code{\link{plotPD}()}
}
\author{
David Carslaw
}
\concept{deweather model plotting functions}
