% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dw.R
\name{predict_dw}
\alias{predict_dw}
\title{Use a deweather model to predict with a new dataset}
\usage{
predict_dw(
  dw,
  newdata = deweather::get_dw_input_data(dw),
  name = deweather::get_dw_pollutant(dw),
  column_bind = FALSE
)
}
\arguments{
\item{dw}{A deweather model created with \code{\link[=build_dw_model]{build_dw_model()}}.}

\item{newdata}{Data set to which to apply the model. If missing the data used
to build the model in the first place will be used.}

\item{name}{The name of the new column.}

\item{column_bind}{If \code{TRUE}, this function will return \code{newdata} with an
additional prediction column bound to it. If \code{FALSE}, return a
single-column data frame.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function is a convenient wrapper around \code{\link[parsnip:predict.model_fit]{parsnip::predict.model_fit()}}
to use a deweather model for prediction. This automatically extracts relevant
parts of the deweather object and creates variables within \code{newdata} using
\code{\link[=append_dw_vars]{append_dw_vars()}} if required.
}
\author{
Jack Davison
}
