% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Prepare data for boosting algorithm}
\usage{
prepData(mydata, add = c("hour", "hour.local", "weekday", "trend", "week",
  "jday", "month"), local.tz = "Europe/London", lag = NULL)
}
\arguments{
\item{mydata}{A data frame to process.}

\item{add}{Names of explanatory variables to include.}

\item{local.tz}{Used if hour needs to be expressed in local
time. This can be useful for situations where the anthropogenic
emissions source is strong and follows local time rather than UTC.}

\item{lag}{Variables(s) to lag. Any variables included here will
add new columns to the data frame. For example \code{lag = "ws"}
with add a new columns \code{lag1ws}. Adding some varaibles here
can improve the explanatory power of the models.}
}
\value{
A data frame with new variables.
}
\description{
Function to prepare data frame for modelling
}
\details{
Info here
}
\author{
David Carslaw
}

