% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dw_met.R
\name{simulate_dw_met}
\alias{simulate_dw_met}
\title{Function to run random meteorological simulations on a deweather model}
\usage{
simulate_dw_met(
  dw,
  newdata = deweather::get_dw_input_data(dw),
  vars = c("ws", "wd", "air_temp"),
  resampling = c("constrained", "all"),
  window_day = 30,
  window_hour = 2,
  n = 200,
  aggregate = TRUE,
  progress = rlang::is_interactive()
)
}
\arguments{
\item{dw}{A deweather model created with \code{\link[=build_dw_model]{build_dw_model()}}.}

\item{newdata}{Data set to which to apply the model. If missing the data used
to build the model in the first place will be used.}

\item{vars}{The variables that should be randomly varied. Note that these
should typically be meteorological variables (e.g., \code{"ws"}, \code{"wd"},
\code{"air_temp"}) and not temporal emission proxies (e.g., \code{"hour"},
\code{"weekday"}, \code{"week"}).}

\item{resampling}{The resampling strategy. One of:
\itemize{
\item \code{"constrained"} (default), meaning that only days of the year close to
the target date are sampled. This option is used in conjunction with
\code{window_day} and \code{window_hour}. For example, a \code{window_day} of \code{30} will
sample +/-30 days of the date.
\item \code{"all"}, meaning all dates are shuffled.
}

The argument for using constrained resampling is that it reduces or
eliminates potentially implausible conditions such as cold temperatures in
summer. It is, however, slower than unconstrained sampling.}

\item{window_day, window_hour}{The day of year (\code{window_day}) and hour of day
(\code{window_hour}) windows to sample within when \code{resampling = "constrained"}.
For example, \code{window_day = 30} samples within +/-30 days of any given date.}

\item{n}{The number of simulations to use.}

\item{aggregate}{By default, all of the simulations will be aggregated into a
single time series. When \code{aggregate = FALSE}, all simulations will be
returned in a single data frame with an \code{.id} column distinguishing between
them.}

\item{progress}{Show a progress bar? Defaults to \code{TRUE} in interactive
sessions.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function performs Monte Carlo simulations to isolate the effect of
emissions changes from meteorological variability in air quality data. It
works by repeatedly shuffling meteorological variables (like wind and air
temperature) while keeping temporal patterns intact, then predicting
pollutant concentrations using a trained deweather model.
}
\author{
David Carslaw
}
