% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqroadside.R
\docType{data}
\name{aqroadside}
\alias{aqroadside}
\title{Example air quality monitoring data for openair}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 149040 rows and 11 columns.
}
\usage{
aqroadside
}
\description{
\code{aqroadside} represents a subset of long-running hourly data from the
Marylebone Road AURN roadside monitoring station in the UK, bound with
meteorological data from the nearby Heathrow airport met station. Five
pollutants (NOx, NO2, Ethane, Isoprene and Benzene) and five meteorlogical
variables (wind speed, wind direction, air temperature, relative humidity,
and cloud height) are provided.
}
\examples{
# basic structure
head(aqroadside)
}
\keyword{datasets}
