% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/boosting.R
\name{buildMod}
\alias{buildMod}
\title{Function to apply meteorological normalisation to.}
\usage{
buildMod(dat, vars = c("ws", "wd"), pollutant = "nox", partial.dep = TRUE,
  B = 100)
}
\arguments{
\item{dat}{Data frame to analyse.}

\item{vars}{Explanatory variables to use}

\item{pollutant}{The name of the variable to apply meteorological
normalisation to.}

\item{partial.dep}{Should the partial dependencies be calculated?}

\item{B}{Number of bootstrap simulations for partial dependence plots.}
}
\value{
A list of stuff
}
\description{
Function to do meteorological normalisation
}
\details{
Info here
}
\author{
David Carslaw
}

