% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMod.R
\name{buildMod}
\alias{buildMod}
\title{Function to apply meteorological normalisation}
\usage{
buildMod(
  input_data,
  vars = c("trend", "ws", "wd", "hour", "weekday", "air_temp"),
  pollutant = "nox",
  sam.size = nrow(input_data),
  n.trees = 200,
  shrinkage = 0.1,
  interaction.depth = 5,
  bag.fraction = 0.5,
  n.minobsinnode = 10,
  cv.folds = 0,
  simulate = FALSE,
  B = 100,
  n.core = 4,
  seed = 123,
  type = "PSOCK"
)
}
\arguments{
\item{input_data}{Data frame to analyse. Must contain a POSIXct field called
\code{date}.}

\item{vars}{Explanatory variables to use. These variables will be used to
build the \code{\link[gbm:gbm]{gbm::gbm()}} model. Note that the model must include a trend
component. Several variables can be automatically calculated (see
\code{\link[=prepData]{prepData()}} for details).}

\item{pollutant}{The name of the variable to apply meteorological
normalisation to.}

\item{sam.size}{The number of random samples to extract from the data for
model building. While it is possible to use the full data set, for data
sets spanning years the model building can take a very long time to run.
Additionally, there will be diminishing returns in terms of model accuracy.
If \code{sam.size} is greater than the number of number of rows of data, the
number of rows of data is used instead.}

\item{n.trees}{Number of trees to fit.}

\item{shrinkage}{A shrinkage parameter applied to each tree in the expansion.
Also known as the learning rate or step-size reduction; \code{0.001} to \code{0.1}
usually work, but a smaller learning rate typically requires more trees.
Default is \code{0.1}.}

\item{interaction.depth}{Integer specifying the maximum depth of each tree
(i.e., the highest level of variable interactions allowed). A value of \code{1}
implies an additive model, a value of \code{2} implies a model with up to 2-way
interactions, etc. Default is \code{5}.}

\item{bag.fraction}{The fraction of the training set observations randomly
selected to propose the next tree in the expansion. This introduces
randomness into the model fit. If \code{bag.fraction < 1} then running the same
model twice will result in similar but different.}

\item{n.minobsinnode}{Integer specifying the minimum number of observations
in the terminal nodes of the trees. Note that this is the actual number of
observations, not the total weight.}

\item{cv.folds}{Number of cross-validation folds to perform. If \code{cv.folds > 1} then \code{\link[gbm:gbm]{gbm::gbm()}}, in addition to the usual fit, will perform a
cross-validation, calculate an estimate of generalization error returned in
\code{cv.error}.}

\item{simulate}{Should the original time series be randomly sampled with
replacement? The default is \code{FALSE}. Setting \code{simulate = TRUE} can be
useful for estimating model uncertainties. In which case models should be
run multiple times with \code{B = 1} and a different value of \code{seed} e.g. \code{seed = runif(1)}.}

\item{B}{Number of bootstrap simulations for partial dependence plots.}

\item{n.core}{Number of cores to use for parallel processing.}

\item{seed}{Random number seed for reproducibility in returned model.}

\item{type}{One of the supported parallelisation types.
See parallel::makeCluster()}
}
\value{
Returns a list including the model, influence data frame and partial
dependence data frame.
}
\description{
This is the main function to apply a \code{\link[gbm:gbm]{gbm::gbm()}} model to a data set.
}
\seealso{
\code{\link[=testMod]{testMod()}} for testing models before they are built.

\code{\link[=metSim]{metSim()}} for using a built model with meteorological simulations.

\code{\link[=plot2Way]{plot2Way()}}, \code{\link[=plotInfluence]{plotInfluence()}} and \code{\link[=plotPD]{plotPD()}} for visualising built
models.
}
\author{
David Carslaw
}
