% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting.R
\name{buildMod}
\alias{buildMod}
\title{Function to apply meteorological normalisation to.}
\usage{
buildMod(dat, vars = c("trend", "ws", "wd", "hour", "weekday", "temp"),
  pollutant = "nox", B = 100, n.core = 4)
}
\arguments{
\item{dat}{Data frame to analyse.}

\item{vars}{Explanatory variables to use.}

\item{pollutant}{The name of the variable to apply meteorological
normalisation to.}

\item{B}{Number of bootstrap simulations for partial dependence plots.}

\item{n.core}{Number of cores to use for parallel processing.}
}
\value{
Returns a list including the model, influence data frame
and partial dependence data frame.
}
\description{
Function to do meteorological normalisation
}
\details{
Info here
}
\author{
David Carslaw
}

