% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diurnals.R
\name{diurnal.gbm}
\alias{diurnal.gbm}
\title{Plot diurnal changes, removing the effect of meteorology}
\usage{
diurnal.gbm(mydata, variables = variables, sub.size = 1,
  pollutant = "nox", test = FALSE, date = c("19/2/2005", "19/2/2007",
  "19/2/2010"), single = FALSE)
}
\arguments{
\item{mydata}{}

\item{variables}{}

\item{sub.size}{}

\item{pollutant}{}

\item{test}{}

\item{date}{}

\item{single}{}
}
\value{
Some data
}
\description{
To add
}
\author{
David Carslaw
}

