% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diurnals.R
\name{diurnalGbm}
\alias{diurnalGbm}
\title{Plot diurnal changes, removing the effect of meteorology}
\usage{
diurnalGbm(dat, vars = c("ws", "wd", "hour", "weekday"), pollutant = "nox",
  date = c("01/01/2012", "31/12/2012", "31/12/2013"), single = FALSE)
}
\arguments{
\item{dat}{A data frame to analyse.}

\item{vars}{The explanatory variables used in the model.}

\item{pollutant}{Name of teh pollutant to apply meteorological
normalisation to.}

\item{date}{A vector of dates, length three. These dates are used
to parition the data into two categories (before/after). The date
format is UK e.g. \code{date = c("19/2/2005", "19/2/2007",
"19/2/2010")}.}

\item{single}{Not used.}
}
\value{
Some data
}
\description{
To add
}
\author{
David Carslaw
}

