% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{gbm.interactions}
\alias{gbm.interactions}
\title{Quantify most important 2-way interactions}
\usage{
gbm.interactions(res)
}
\arguments{
\item{res}{Model object from running \code{buildMod}.}
}
\value{
Interaction information.
}
\description{
Two-way interactions
}
\author{
David Carslaw
}

