% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{gbmInf}
\alias{gbmInf}
\title{Influence plot}
\usage{
gbmInf(dat)
}
\arguments{
\item{dat}{Model object from running \code{buildMod}}
}
\value{
Plot
}
\description{
Plot a gbm influence plot
}
\author{
David Carslaw
}

