% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dw.R
\name{get_tdw_pollutant}
\alias{get_tdw_pollutant}
\alias{get_tdw_vars}
\alias{get_tdw_engine}
\alias{get_tdw_best_params}
\alias{get_tdw_tuning_metrics}
\alias{get_tdw_testing_metrics}
\alias{get_tdw_testing_data}
\title{Getters for various deweather tuning object features}
\usage{
get_tdw_pollutant(tdw)

get_tdw_vars(tdw)

get_tdw_engine(tdw)

get_tdw_best_params(tdw, param = NULL)

get_tdw_tuning_metrics(tdw, metric = NULL)

get_tdw_testing_metrics(tdw, metric = NULL)

get_tdw_testing_data(tdw)
}
\arguments{
\item{tdw}{A deweather tuning object created with \code{\link[=tune_dw_model]{tune_dw_model()}}.}

\item{param}{For \code{\link[=get_dw_params]{get_dw_params()}} and \code{\link[=get_tdw_best_params]{get_tdw_best_params()}}. The default
(\code{NULL}) returns a list of model parameters. \code{param} will return one
specific parameter as a character vector.}

\item{metric}{For \code{\link[=get_tdw_tuning_metrics]{get_tdw_tuning_metrics()}} and \code{\link[=get_tdw_testing_metrics]{get_tdw_testing_metrics()}}.
The default (\code{NULL}) returns a complete set of model parameters. \code{metric}
will return one specific parameter. \code{metric} must be one of the
\code{\link[openair:aqStats]{openair::aqStats()}} metrics for \code{\link[=get_tdw_tuning_metrics]{get_tdw_tuning_metrics()}} and one of
\code{"rmse"} or \code{"rsq"} for \code{\link[=get_tdw_testing_metrics]{get_tdw_testing_metrics()}}.}
}
\value{
Typically a character vector, except:
\itemize{
\item \code{\link[=get_tdw_testing_metrics]{get_tdw_testing_metrics()}}: a list
\item \code{\link[=get_tdw_best_params]{get_tdw_best_params()}}: a list, unless \code{param} is set.
\item \code{\link[=get_tdw_testing_data]{get_tdw_testing_data()}}: a \code{data.frame}
}
}
\description{
\code{deweather} provides multiple 'getter' functions for extracting relevant
model features from a deweather model and/or tuning objects. These are a
useful convenience, particularly in conjunction with R's \link[=pipeOp]{pipe}
operator (\verb{|>}).
}
\seealso{
Other Object 'Getter' Functions: 
\code{\link{get_dw_pollutant}()}
}
\concept{Object 'Getter' Functions}
