% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metSim.R
\name{metSim}
\alias{metSim}
\title{Run random meteorology on a gbm model}
\usage{
metSim(dat, newdata, metVars = c("ws", "wd", "temp"), n.core = 4, B = 200)
}
\arguments{
\item{dat}{Model object from running \code{buildMod}.}

\item{newdata}{Model object from running \code{buildMod}.}

\item{metVars}{Teh variables that should be randomly varied.}

\item{n.core}{Number of cores to use}

\item{B}{Number of simulations}
}
\value{
To add
}
\description{
Function to run random meteorological simulations on a gbm model
}
\author{
David Carslaw
}

