% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metSim.R
\name{metSim}
\alias{metSim}
\title{Function to run random meteorological simulations on a gbm model}
\usage{
metSim(
  dw_model,
  newdata,
  metVars = c("ws", "wd", "air_temp"),
  n.core = 4,
  B = 200,
  progress = TRUE
)
}
\arguments{
\item{dw_model}{Model object from running \code{\link[=buildMod]{buildMod()}}.}

\item{newdata}{Data set to which to apply the model. If missing the data used
to build the model in the first place will be used.}

\item{metVars}{The variables that should be randomly varied.}

\item{n.core}{Number of cores to use.}

\item{B}{Number of simulations}

\item{progress}{When using multiple cores, show a progress indicator for
bootstrap simulations?}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Function to run random meteorological simulations on a gbm model
}
\seealso{
\code{\link[=buildMod]{buildMod()}} to build a gbm model
}
\author{
David Carslaw
}
