% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metSim.R
\name{metSim}
\alias{metSim}
\title{Run random meteorology on a gbm model}
\usage{
metSim(dw_model, newdata, metVars = c("ws", "wd", "temp"), n.core = 4, B = 200)
}
\arguments{
\item{dw_model}{Model object from running \code{buildMod}.}

\item{newdata}{Data set to which to apply the model. If missing
the data used to build the model in the first place will be
used.}

\item{metVars}{The variables that should be randomly varied.}

\item{n.core}{Number of cores to use.}

\item{B}{Number of simulations}
}
\value{
To add
}
\description{
Function to run random meteorological simulations on a gbm model
}
\author{
David Carslaw
}
