% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partialDep.R
\name{plot2Way}
\alias{plot2Way}
\title{Plot two-way interactions from gbm model}
\usage{
plot2Way(dat, variable = c("ws", "temp"), ...)
}
\arguments{
\item{dat}{Model object from running \code{buildMod}.}

\item{variable}{The variables to plot. Must be of length two
e.g. \code{variables = c("ws", "wd"}.}

\item{...}{other arguments to send to \code{openair} \code{scatterPlot}.}
}
\value{
To add
}
\description{
Two-way intercation plots
}
\details{
To add
}
\author{
David Carslaw
}

