% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDep.R
\name{plot2Way}
\alias{plot2Way}
\title{Plot two-way interactions from gbm model}
\usage{
plot2Way(dw_model, variable = c("ws", "temp"), res = 100, exlude = TRUE,
  cols = "default", dist = 0.05, ...)
}
\arguments{
\item{dw_model}{Model object from running \code{buildMod}.}

\item{variable}{The variables to plot. Must be of length two
e.g. \code{variables = c("ws", "wd"}.}

\item{res}{Resolution in x-y i.e. number of points in each dimension.}

\item{exlude}{Should surfaces exlude predictions too far from
original data? The default is \code{TRUE}.}

\item{cols}{Colours to be used for plotting. Options include
\dQuote{default}, \dQuote{increment}, \dQuote{heat}, \dQuote{jet}
and user defined. For user defined the user can supply a list of
colour names recognised by R (type \code{colours()} to see the
full list). An example would be \code{cols = c("yellow", "green",
"blue")}}

\item{dist}{When plotting surfaces, \code{dist} controls how far from the
original data the predictions should be made. See
\code{exclude.too.far} from the \code{mgcv} package. Data are
first transformed to a unit square. Values should be between 0 and
1.}

\item{...}{Other arguments to be passed for plotting.}
}
\value{
To add
}
\description{
Two-way interaction plots
}
\details{
To add
}
\author{
David Carslaw
}
