% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dw_importance.R
\name{plot_dw_importance}
\alias{plot_dw_importance}
\title{Visualise deweather model feature importance}
\usage{
plot_dw_importance(dw, aggregate_factors = FALSE, sort = TRUE, cols = "tol")
}
\arguments{
\item{dw}{A deweather model created with \code{\link[=build_dw_model]{build_dw_model()}}.}

\item{aggregate_factors}{Defaults to \code{FALSE}. If \code{TRUE}, the importance of
factor inputs (e.g., Weekday) will be summed into a single variable. This
only applies to certain engines (e.g., \code{"xgboost"}) which report factor
importance as disaggregate features.}

\item{sort}{If \code{TRUE}, the default, features will be sorted by their
importance. If \code{FALSE}, they will be sorted alphabetically. In
\code{\link[=plot_dw_importance]{plot_dw_importance()}} this will change the ordering of the y-axis, whereas
in \code{\link[=get_dw_importance]{get_dw_importance()}} it will change whether \code{var} is returned as a
factor or character data type.}

\item{cols}{Colours to use for plotting. See \code{\link[openair:openColours]{openair::openColours()}}.}
}
\value{
a \link[ggplot2:ggplot2-package]{ggplot2} figure
}
\description{
Visualise the feature importance (\% Gain for boosted tree models) for each
variable of a deweather model, with some customisation.
}
