% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dw_partial_1d.R
\name{plot_dw_partial_1d}
\alias{plot_dw_partial_1d}
\title{Create partial dependence plots for deweather models}
\usage{
plot_dw_partial_1d(
  dw,
  vars = NULL,
  intervals = 40L,
  group = NULL,
  group_intervals = 3L,
  show_conf_int = TRUE,
  n = NULL,
  prop = 0.1,
  cols = "tol",
  radial_wd = TRUE,
  ncol = NULL,
  nrow = NULL,
  plot = TRUE,
  progress = rlang::is_interactive()
)
}
\arguments{
\item{dw}{A deweather model created with \code{\link[=build_dw_model]{build_dw_model()}}.}

\item{vars}{Character. The name of the variable(s) to plot. Must be one of
the variables used in the model. If \code{NULL}, all variables will be plotted
in order of importance.}

\item{intervals}{The number of points for the partial dependence profile.}

\item{group}{Optional grouping variable to show separate profiles for
different levels of another predictor. Must be one of the variables used in
the model. Default is \code{NULL} (no grouping).}

\item{group_intervals}{The number of bins when the \code{group} variable is
numeric.}

\item{show_conf_int}{Should the bootstrapped 95\% confidence interval be
shown? In \code{\link[=plot_dw_partial_1d]{plot_dw_partial_1d()}} these are shown using transparent ribbons
(for numeric variables) and rectangles (for categorical variables).}

\item{n}{The number of observations to use for calculating the partial
dependence profile. If \code{NULL} (default), uses \code{prop} to determine the
sample size.}

\item{prop}{The proportion of input data to use for calculating the partial
dependence profile, between 0 and 1. Default is \code{0.1} (10\% of data).
Ignored if \code{n} is specified.}

\item{cols}{Colours to use for plotting. See \code{\link[openair:openColours]{openair::openColours()}}.}

\item{radial_wd}{Should the \code{"wd"} (wind direction) variable be plotted on a
radial axis? This can enhance interpretability, but makes it inconsistent
with other variables which are plotted on cartesian coordinates. Defaults
to \code{TRUE}.}

\item{ncol, nrow}{When more than one \code{vars} is defined, \code{ncol} and \code{nrow}
define the dimensions of the grid to create. Setting both to be \code{NULL}
creates a roughly square grid.}

\item{plot}{When \code{FALSE}, return a list of plot data instead of a plot.}

\item{progress}{Show a progress bar? Defaults to \code{TRUE} in interactive
sessions.}
}
\value{
A \code{ggplot2} object showing the partial dependence plot. If multiple
\code{vars} are specified, a \code{patchwork} assembly of plots will be returned. If
\code{plot = FALSE}, a named list of plot data will be returned instead.
}
\description{
Generates partial dependence plots to visualize the relationship between
predictor variables and model predictions. These plots show how the predicted
pollutant concentration changes as a function of one variable while averaging
over the effects of all other variables.
}
