% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_trend.R
\name{plot_sim_trend}
\alias{plot_sim_trend}
\title{Plot a simulated 'deweathered' trend, optionally with its input data}
\usage{
plot_sim_trend(
  sim,
  dw = NULL,
  avg.time = NULL,
  names = c("Met Simulation", "Original Data"),
  cols = "tol",
  ...,
  .plot_engine = c("ggplot2", "plotly")
)
}
\arguments{
\item{sim}{The output of \code{\link[=simulate_dw_met]{simulate_dw_met()}}; a \code{data.frame} with first
column \code{date} and second column a numeric pollutant.}

\item{dw}{Optionally, the input model used to construct \code{sim}. If provided,
the original timeseries will be plotted alongside the simulated trend for
easy comparison.}

\item{avg.time}{Passed to \code{\link[openair:timeAverage]{openair::timeAverage()}}.}

\item{names}{A character vector of length two, used to label the simulated
dataset and the original dataset.}

\item{cols}{Colours to use for plotting. See \code{\link[openair:openColours]{openair::openColours()}}.}

\item{...}{Not currently used.}

\item{.plot_engine}{The plotting engine to use. One of \code{"ggplot2"}, which
returns a static plot, or \code{"plotly"}, which returns a dynamic HTML plot.}
}
\description{
This function conveniently plots the trend from a simulated deweathered time
series, with the option to overlay the trend from the original input data.
The data can also be averaged over specified time intervals for clearer
visualisation.
}
\author{
Jack Davison
}
