% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tdw_testing_scatter.R
\name{plot_tdw_testing_scatter}
\alias{plot_tdw_testing_scatter}
\title{Plot Observed vs Modelled Scatter using the 'best parameters' from
\code{\link[=tune_dw_model]{tune_dw_model()}}}
\usage{
plot_tdw_testing_scatter(
  tdw,
  method = c("scatter", "bin", "hexbin"),
  group = NULL,
  bins = 50L,
  show_ablines = TRUE,
  show_params = TRUE,
  cols = "viridis",
  cols_ablines = c("black", "grey50"),
  ...,
  .plot = TRUE,
  .plot_engine = c("ggplot2", "plotly")
)
}
\arguments{
\item{tdw}{A deweather tuning object created with \code{\link[=tune_dw_model]{tune_dw_model()}}.}

\item{method}{One of \code{"scatter"}, \code{"bin"} or \code{"hexbin"}.}

\item{group}{A variable (one of the initial modelling parameters) to colour
the scatter plot by. Only used when \code{method = "scatter"}. This could be
useful to determine where the model is working most or least effectively,
or to identify other patterns in the data.}

\item{bins}{The number of bins to use when \code{method = "bin"} or \code{method = "hexbin"}.}

\item{show_ablines}{Show 1:1, 2:1 and 1:2 lines to assist with model
evaluation? Lines will appear beneath the "scatter" \code{method} and above
either of the "bin" \code{method}s.}

\item{show_params}{Show an annotation of model parameters in the top-left
corner of the scatter plot?}

\item{cols}{Colours to use for plotting. See \code{\link[openair:openColours]{openair::openColours()}}.}

\item{cols_ablines}{Colours to use for the diagonal lines, if \code{show_ablines = TRUE}. The the first colour is used for the 1:1 line, and the second for
the 2:1 and 1:2 lines. Passed to \code{\link[openair:openColours]{openair::openColours()}}.}

\item{...}{Not currently used.}

\item{.plot}{When \code{FALSE}, return a \code{data.frame} of plot data instead of a
plot.}

\item{.plot_engine}{The plotting engine to use. One of \code{"ggplot2"}, which
returns a static plot, or \code{"plotly"}, which returns a dynamic HTML plot.}
}
\description{
\code{\link[=tune_dw_model]{tune_dw_model()}} determines a 'best' set of parameters automatically and
models some 'final' predictions using a reserved testing dataset to evaluate
the model. This function produces a scatter (or varient)
}
\seealso{
Other Model Tuning Functions: 
\code{\link{plot_tdw_tuning_metrics}()},
\code{\link{tune_dw_model}()}
}
\concept{Model Tuning Functions}
