% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tdw_tuning_metrics.R
\name{plot_tdw_tuning_metrics}
\alias{plot_tdw_tuning_metrics}
\title{Plot Tuning Metrics from \code{\link[=tune_dw_model]{tune_dw_model()}}}
\usage{
plot_tdw_tuning_metrics(
  tdw,
  x = NULL,
  group = NULL,
  facet = NULL,
  show_std_err = TRUE,
  cols = "tol",
  ...,
  .plot_engine = c("ggplot2", "plotly")
)
}
\arguments{
\item{tdw}{A deweather tuning object created with \code{\link[=tune_dw_model]{tune_dw_model()}}.}

\item{x}{The tuned parameter to plot on the x-axis. If not selected, the
first parameter in the \code{metrics} dataset will be chosen.}

\item{group, facet}{Additional tuned parameters other than \code{x}, used to
further control the plot. \code{group} colours the plot by another parameter,
and \code{facet} splits the diagram into additional panels. Neither \code{group} nor
\code{facet} can be the same parameter as \code{x}.}

\item{show_std_err}{Show the standard error using error bars?}

\item{cols}{Colours to use for plotting. See \code{\link[openair:openColours]{openair::openColours()}}.}

\item{...}{Not currently used.}

\item{.plot_engine}{The plotting engine to use. One of \code{"ggplot2"}, which
returns a static plot, or \code{"plotly"}, which returns a dynamic HTML plot.}
}
\description{
This function creates a plot of the tuning metrics from a \code{TuneDeweather}
object created using \code{\link[=tune_dw_model]{tune_dw_model()}}. It visualises how different
hyperparameter values affect model performance (RMSE and RSQ). This allows
for the 'best' parameters to be refined through visual inspection. This plot
is likely most effective with between 1 and 3 simultaneously tuned
parameters; any more will impede plot interpretation.
}
\seealso{
Other Model Tuning Functions: 
\code{\link{plot_tdw_testing_scatter}()},
\code{\link{tune_dw_model}()}
}
\author{
Jack Davison
}
\concept{Model Tuning Functions}
