% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDep.R
\name{plotAllPD}
\alias{plotAllPD}
\title{Plot all partial dependencies}
\usage{
plotAllPD(dw_model, ylim = NULL, nrow = NULL, ...)
}
\arguments{
\item{dw_model}{Model object from running \code{buildMod}.}

\item{ylim}{y-axis label. Will default to pollutant name.}

\item{nrow}{Number of rows for the plots.}

\item{...}{extra plotting arguments.}
}
\value{
A plot
}
\description{
Plot all partial dependencies
}
\author{
David Carslaw
}
