% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDep.R
\name{plotAllPD}
\alias{plotAllPD}
\title{Plot all partial dependencies}
\usage{
plotAllPD(dat, ylim = NULL, nrow = NULL, ylab = NULL, ...)
}
\arguments{
\item{dat}{Model object from running \code{buildMod}.}

\item{ylim}{y-axis label. Will default to pollutant name.}

\item{nrow}{Number of rows for th eplots.}

\item{ylab}{The y-axis label to over-ride default.}

\item{...}{extra plotting arguments.}
}
\value{
A plot
}
\description{
Function to plot all partial dependencies
}
\details{
.
}
\author{
David Carslaw
}

