% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInfluence.R
\name{plotInfluence}
\alias{plotInfluence}
\title{Plot a GBM influence plot}
\usage{
plotInfluence(dw_model, col = "grey30", sort = TRUE)
}
\arguments{
\item{dw_model}{Model object from running \code{\link[=buildMod]{buildMod()}}.}

\item{col}{Colour to use to use to colour the bars. Alternatively, users can
provide \code{"var"} which will colour each bar differently, or \code{"mean"} which
will colour each bar by its mean relative variable influence.}

\item{sort}{Sort the variables by their relative variable influences?
Defaults to \code{TRUE}. \code{FALSE} displays them alphabetically.}
}
\value{
Plot
}
\description{
Plot a GBM influence plot
}
\seealso{
Other deweather model plotting functions: 
\code{\link{plot2Way}()},
\code{\link{plotPD}()}
}
\author{
David Carslaw
}
\concept{deweather model plotting functions}
