% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPD.R
\name{plotPD}
\alias{plotPD}
\title{Function to plot partial dependence plots with bootstrap uncertainties}
\usage{
plotPD(
  dw_model,
  variable,
  ylim = NULL,
  plotit = TRUE,
  intervals = 40,
  auto.text = TRUE,
  ...
)
}
\arguments{
\item{dw_model}{Model object from running \code{\link[=buildMod]{buildMod()}}.}

\item{variable}{The variable to plot.}

\item{ylim}{user-specified \code{ylim}.}

\item{plotit}{Should a plot be produced?}

\item{intervals}{Number of intervals to to calculate partial dependence over.}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If \code{TRUE}
titles and axis labels will automatically try and format pollutant names
and units properly e.g.  by subscripting the `2' in NO2.}

\item{...}{other arguments for plotting.}
}
\value{
Invisibly returns a list containing the plot and the data used to
make the plot for \code{\link[=plotPD]{plotPD()}} and \code{\link[=plot2Way]{plot2Way()}} to allow post
processing, e.g., modifying the \code{ggplot}.
}
\description{
Function to plot partial dependence plots with bootstrap uncertainties
}
\author{
David Carslaw
}
