% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partialDep.R
\name{plotPD}
\alias{plotPD}
\title{Partial dependence plots with uncertainties.}
\usage{
plotPD(dat, variable, ylim = NULL, ...)
}
\arguments{
\item{dat}{Model object from running \code{buildMod}.}

\item{variable}{The variable to plot.}

\item{ylim}{user-specified \code{ylim}.}

\item{...}{other arguments for plotting.}
}
\value{
A plot
}
\description{
Function to plot partial dependence plots with bootstrap uncertainties
}
\author{
David Carslaw
}

