% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialDep.R
\name{plotPD}
\alias{plotPD}
\title{Partial dependence plots with uncertainties.}
\usage{
plotPD(dat, variable, ylim = NULL, plotit = TRUE, auto.text = TRUE,
  ...)
}
\arguments{
\item{dat}{Model object from running \code{buildMod}.}

\item{variable}{The variable to plot.}

\item{ylim}{user-specified \code{ylim}.}

\item{plotit}{Should a plot be produced?}

\item{auto.text}{Either \code{TRUE} (default) or \code{FALSE}. If
\code{TRUE} titles and axis labels will automatically try and
format pollutant names and units properly e.g.  by subscripting
the `2' in NO2.}

\item{...}{other arguments for plotting.}
}
\value{
A plot
}
\description{
Function to plot partial dependence plots with bootstrap uncertainties
}
\author{
David Carslaw
}
