% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Function to prepare data frame for modelling}
\usage{
prepData(
  mydata,
  add = c("hour", "hour.local", "weekday", "trend", "week", "jday", "month"),
  local.tz = "Europe/London",
  lag = NULL
)
}
\arguments{
\item{mydata}{A data frame to process.}

\item{add}{Names of explanatory variables to include.}

\item{local.tz}{Used if hour needs to be expressed in local time. This can be
useful for situations where the anthropogenic emissions source is strong
and follows local time rather than UTC.}

\item{lag}{Variables(s) to lag. Any variables included here will add new
columns to the data frame. For example \code{lag = "ws"} with add a new
columns \code{lag1ws}. Adding some variables here can improve the
explanatory power of the models. Variables are lagged by one unit of time.}
}
\value{
A data frame with new variables.
}
\description{
This function takes a data frame that contains a field \code{date} and other
variables and adds other common variables needed by the modelling functions.
This function is run automatically by \code{\link[=buildMod]{buildMod()}} but can be used separately
for further analysis. These variables include:
\itemize{
\item \strong{hour} - The hour of the day from 0 to 23.
\item \strong{hour.local} - The hour in the local time zone. Note that the local time
zone will need to be supplied (see \code{local.tz}). The purpose of using
local time rather than UTC is that emissions can vary more strongly by local
time rather than UTC.
\item \strong{weekday} - The day of the week.
\item \strong{trend} - The trend is calculated as a decimal year.
\item \strong{week} - The week of the year. Useful for taking account of long-term
seasonal variations.
\item \strong{jday} - The Julian Day number.
\item \strong{month} - month of the year. Useful for taking account of long-term
seasonal variations.
}
}
\author{
David Carslaw
}
