% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{road_data}
\alias{road_data}
\title{Example data for deweather}
\format{
Data frame with 65533 observations (rows) on the following 10
variables: \describe{ \item{list("date")}{Observation date/time stamp in
year-month-day hour:minute:second format (POSIXct). } \item{list("nox",
"no2", "ethane", "isoprene", "benzene")}{Pollutant concentrations.}
\item{list("pm25")}{Particulate PM2.5 fraction measurement, in ug/m3, as a
numeric vector.}\item{list("ws")}{Wind speed, in m/s, as numeric
vector.}\item{list("wd")}{Wind direction, in degrees from North, as a
numeric vector.}\item{list("air_temp")}{Air temperature, in degrees
Celcius.}\item{list("RH")}{Percentage relative humidity.}}
}
\source{
\code{deweather} was compiled from data archived in the London Air
Quality Archive.  See \url{http://www.londonair.org.uk} for site details.

The same data is also provide in \code{'.csv'} format via the openair
project web site \url{http://www.openair-project.org}.
}
\usage{
road_data
}
\description{
To complete.
}
\details{
\code{road_data} is supplied with the \code{deweather} package as an example
dataset for use with documented examples.
}
\note{
\code{deweather} functions generally require data frames with a field
"date" that can be in either \code{POSIXct} or \code{Date} format but
should be GMT time zone. This can be hourly data or higher resolution data.
}
\examples{

# basic structure
head(road_data)
}
\seealso{
Other built-in datasets: 
\code{\link{dat}}
}
\concept{built-in datasets}
\keyword{datasets}
