% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMod.R
\name{testMod}
\alias{testMod}
\title{Function to test different meteorological normalisation models.}
\usage{
testMod(
  dat,
  vars = c("trend", "ws", "wd", "hour", "weekday", "temp"),
  pollutant = "nox",
  train.frac = 0.8,
  n.trees = 200,
  seed = 123
)
}
\arguments{
\item{dat}{Data frame to analyse.}

\item{vars}{Explanatory variables to use.}

\item{pollutant}{The name of the variable to apply meteorological
normalisation to.}

\item{train.frac}{Fraction of data to train a model on. The model
is tested against the withheld 0.2 proportion.}

\item{n.trees}{Number of trees to use.}

\item{seed}{Random number seed for reproducibility in returned model.}
}
\value{
Returns to be added.
}
\description{
Function to test different meteorological normalisation models.
}
\author{
David Carslaw
}
