% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_dw_model.R
\name{tune_dw_model}
\alias{tune_dw_model}
\title{Tune a deweather model}
\usage{
tune_dw_model(
  data,
  pollutant,
  vars = c("trend", "ws", "wd", "hour", "weekday", "air_temp"),
  tree_depth = 5,
  trees = 200L,
  learn_rate = 0.1,
  mtry = NULL,
  min_n = 10L,
  loss_reduction = 0,
  sample_size = 1L,
  stop_iter = 190L,
  engine = c("xgboost", "lightgbm"),
  split_prop = 3/4,
  grid_levels = 5,
  v_partitions = 10
)
}
\arguments{
\item{data}{An input \code{data.frame} containing one pollutant column (defined
using \code{pollutant}) and a collection of feature columns (defined using
\code{vars}).}

\item{pollutant}{The name of the column (likely a pollutant) in \code{data} to
predict.}

\item{vars}{The name of the columns in \code{data} to use as model features -
i.e., to predict the values in the \code{pollutant} column. Any character
columns will be coerced to factors. \code{"hour"}, \code{"weekday"}, \code{"trend"},
\code{"yday"}, \code{"week"}, and \code{"month"} are special terms and will be passed to
\code{\link[=append_dw_vars]{append_dw_vars()}} if not present in \code{names(data)}.}

\item{tree_depth, trees, learn_rate, mtry, min_n, loss_reduction, sample_size, stop_iter}{If length 1, these parameters will be fixed. If length \code{2}, the parameter
will be tuned within the range defined between the first and last value. For
example, if \code{tree_depth = c(1, 5)} and \code{grid_levels = 3}, tree depths of \code{1},
\code{3}, and \code{5} will be tested.}

\item{engine}{A single character string specifying what computational engine
to use for fitting.}

\item{split_prop}{The proportion of data to be retained for
modeling/analysis. Passed to the \code{prop} argument of
\code{\link[rsample:initial_split]{rsample::initial_split()}}.}

\item{grid_levels}{An integer for the number of values of each parameter to
use to make the regular grid. Passed to the \code{levels} argument of
\code{\link[dials:grid_regular]{dials::grid_regular()}}.}

\item{v_partitions}{The number of partitions of the data set to use for
v-fold cross-validation. Passed to the \code{v} argument of
\code{\link[rsample:vfold_cv]{rsample::vfold_cv()}}.}
}
\description{
This function performs hyperparameter tuning for a gradient boosting model
used in deweathering air pollution data. It uses cross-validation to find
optimal hyperparameters and returns the best performing model along with
performance metrics and visualizations. Parallel processing (e.g., through
the \code{mirai} package) is recommended to speed up tuning - see
\url{https://tune.tidymodels.org/articles/extras/optimizations.html#parallel-processing}.
}
\details{
The function performs the following steps:
\itemize{
\item Removes rows with missing values in the pollutant or predictor variables
\item Splits data into training and testing sets
\item Creates a tuning grid for any parameters specified as ranges
\item Performs grid search with cross-validation to find optimal hyperparameters
\item Fits a final model using the best hyperparameters
\item Generates predictions and performance metrics

At least one hyperparameter must be specified as a range (vector of length
2) for tuning to occur. Single values are treated as fixed parameters.
}
}
\author{
Jack Davison
}
